+++ ballerina/client.bal
@@ -34,11 +34,11 @@ public isolated client class Client {
     public isolated function init(string url, string user, string password,
         Options? options = (), sql:ConnectionPool? connectionPool = ()) returns sql:Error? {
         ClientConfiguration clientConf = {
-            url: url,
-            user: user,
-            password: password,
-            options: options,
-            connectionPool: connectionPool
+            URL,
+            user,
+            password,
+            options,
+            connectionPool
         };
         return createClient(self, clientConf, sql:getGlobalConnectionPool());
     }
@@ -50,7 +50,7 @@ public isolated client class Client {
     # + return - Stream of records in the type of rowType
     remote isolated function query(sql:ParameterizedQuery sqlQuery, typedesc<record {}> rowType = <>)
     returns stream<rowType, sql:Error?> = @java:Method {
-        'class: "io.ballerina.stdlib.java.jdbc.nativeimpl.QueryProcessor",
+        'class: "QueryProcessor",
         name: "nativeQuery"
     } external;
 
@@ -63,7 +63,7 @@ public isolated client class Client {
     # + return - Result in the returnType type or an sql:Error
     remote isolated function queryRow(sql:ParameterizedQuery sqlQuery, typedesc<anydata> returnType = <>)
     returns returnType|sql:Error = @java:Method {
-        'class: "io.ballerina.stdlib.java.jdbc.nativeimpl.QueryProcessor",
+        'class: "QueryProcessor",
         name: "nativeQueryRow"
     } external;
 
@@ -73,7 +73,7 @@ public isolated client class Client {
     # + return - Metadata of the query execution as an sql:ExecutionResult or an sql:Error
     remote isolated function execute(sql:ParameterizedQuery sqlQuery)
     returns sql:ExecutionResult|sql:Error = @java:Method {
-        'class: "io.ballerina.stdlib.java.jdbc.nativeimpl.ExecuteProcessor",
+        'class: "ExecuteProcessor",
         name: "nativeExecute"
     } external;
 
@@ -98,7 +98,7 @@ public isolated client class Client {
     # + return - Summary of the execution and results are returned in an `sql:ProcedureCallResult`, or an `sql:Error`
     remote isolated function call(sql:ParameterizedCallQuery sqlQuery, typedesc<record {}>[] rowTypes = [])
     returns sql:ProcedureCallResult|sql:Error = @java:Method {
-        'class: "io.ballerina.stdlib.java.jdbc.nativeimpl.CallProcessor",
+        'class: "CallProcessor",
         name: "nativeCall"
     } external;
 
@@ -106,7 +106,7 @@ public isolated client class Client {
     #
     # + return - Possible error when closing the client
     public isolated function close() returns sql:Error? = @java:Method {
-        'class: "io.ballerina.stdlib.java.jdbc.nativeimpl.ClientProcessor",
+        'class: "ClientProcessor",
         name: "close"
     } external;
 }
@@ -138,10 +138,10 @@ type ClientConfiguration record {|
 
 isolated function createClient(Client jdbcClient, ClientConfiguration clientConf,
         sql:ConnectionPool globalConnPool) returns sql:Error? = @java:Method {
-    'class: "io.ballerina.stdlib.java.jdbc.nativeimpl.ClientProcessor"
+    'class: "ClientProcessor"
 } external;
 
 isolated function nativeBatchExecute(Client sqlClient, string[]|sql:ParameterizedQuery[] sqlQueries)
 returns sql:ExecutionResult[]|sql:Error = @java:Method {
-    'class: "io.ballerina.stdlib.java.jdbc.nativeimpl.ExecuteProcessor"
+    'class: "ExecuteProcessor"
 } external;
