+++ build-config/resources/Ballerina.toml
@@ -3,7 +3,7 @@ org = "ballerinax"
 name = "aws.redshift"
 version = "@toml.version@"
 authors = ["Ballerina"]
-keywords = ["Data Warehouse", "Columnar Storage", "Cost/Paid"]
+keywords = ["Data Warehouse", "Columnar Storage", "Cost/Paid", "vendor/aws"]
 repository = "https://github.com/ballerina-platform/module-ballerinax-aws.redshift"
 icon = "icon.png"
 license = ["Apache-2.0"]
