+++ build.gradle
@@ -29,8 +29,7 @@ description = 'Ballerina - AWS Redshift'
 allprojects {
     group = project.group
     version = project.version
-
-    apply plugin: 'jacoco'
+    
     apply plugin: 'maven-publish'
 
     repositories {
