+++ gradle.properties
@@ -1,5 +1,5 @@
-group=io.ballerinax
-version=0.1.0
+group=io.ballerina.lib
+version=0.1.0-SNAPSHOT
 
 puppycrawlCheckstyleVersion=10.12.1
 checkstyleToolVersion=10.12.1
@@ -9,5 +9,4 @@ underCouchDownloadVersion=5.4.0
 researchgateReleaseVersion=2.8.0
 testngVersion=7.6.1
 
-ballerinaLangVersion=2201.8.0-20230830-220400-8a7556d8
-stdlibSqlVersion=1.11.0
+ballerinaLangVersion=2201.8.0
