+++ native/src/main/java/io/ballerina/lib/aws/redshift/nativeimpl/ClientProcessor.java
@@ -16,13 +16,13 @@
  * under the License.
  */
 
-package io.ballerina.stdlib.java.jdbc.nativeimpl;
+package io.ballerina.lib.aws.redshift.nativeimpl;
 
 import io.ballerina.runtime.api.creators.ValueCreator;
 import io.ballerina.runtime.api.values.BMap;
 import io.ballerina.runtime.api.values.BObject;
 import io.ballerina.runtime.api.values.BString;
-import io.ballerina.stdlib.java.jdbc.Constants;
+import io.ballerina.lib.aws.redshift.Constants;
 import io.ballerina.stdlib.sql.datasource.SQLDatasource;
 import io.ballerina.stdlib.sql.utils.ErrorGenerator;
